Moved type to constants
