+++ packages/zendesk-adapter/src/constants.ts
@@ -75,3 +75,4 @@ export const TRIGGER_TYPE_NAME = 'trigger'
 export const LOCALE_TYPE_NAME = 'locale'
 export const TICKET_STATUS_CUSTOM_STATUS_TYPE_NAME = 'custom_status'
 export const DEFLECTION_TYPE = 'deflection'
+export const VIEW_TYPE_NAME = 'view'
